pyproject.toml
[tool.poetry]
name = "job-is"
version = "0.0.1"
description = "job-is backend"
authors = ["An jaegyeong <anjaekk@gmail.com>"]
readme = "README.md"
packages = [{include = "job_is"}]

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.7.0"
attrs = "23.1.0"
build = "0.10.0"
cachecontrol = "0.12.14"
certifi = "2023.5.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
cleo = "2.0.1"
click = "8.1.3"
crashtest = "0.4.1"
distlib = "0.3.6"
dulwich = "0.21.5"
exceptiongroup = "1.1.1"
fastapi = "0.96.0"
filelock = "3.12.0"
h11 = "0.14.0"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "6.6.0"
installer = "0.7.0"
jaraco-classes = "3.2.3"
jsonschema = "4.17.3"
keyring = "23.13.1"
lockfile = "0.12.2"
more-itertools = "9.1.0"
msgpack = "1.0.5"
packaging = "21.3"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.5.1"
poetry = "1.5.1"
poetry-core = "1.6.1"
poetry-plugin-export = "1.4.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pydantic = "1.10.8"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
rapidfuzz = "2.15.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.0.post1"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.27.0"
tomli = "2.0.1"
tomlkit = "0.11.8"
trove-classifiers = "2023.5.24"
typing-extensions = "4.6.3"
urllib3 = "1.26.16"
uvicorn = "0.22.0"
virtualenv = "20.23.0"
webencodings = "0.5.1"
xattr = "0.10.1"
zipp = "3.15.0"
sqlalchemy = "2.0.15"
async-timeout = "4.0.2"
deprecated = "1.2.13"
envoy = "0.0.3"
greenlet = "2.0.2"
pymysql = "1.0.3"
pyparsing = "3.0.9"
redis = "4.3.4"
wrapt = "1.14.1"
alembic = "1.11.1"
mako = "1.2.4"
markupsafe = "2.1.3"
mysqlclient = "2.1.1"
cryptography = "^41.0.1"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
python-dateutil = "^2.8.2"
fastapi-pagination = "^0.12.4"
httpx = "^0.24.1"
factory-boy = "^3.2.1"
pytz = "^2023.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
